refactor: 에러 네이밍 수정
